+++ test/it/netutils/httpstatus_test.sh
@@ -1,0 +1,3 @@
+TestHttpStatusSearch() {
+    http-status-code search 404
+}