+++ test/it/securityutils/pwgen_test.sh
@@ -1,0 +1,3 @@
+TestPwgenCount() {
+    pwgen -n 3 -l 8 | wc -l | tr -d ' '
+}