+++ test/it/securityutils/pwscore_test.sh
@@ -1,0 +1,3 @@
+TestPwscoreCommon() {
+    pwscore password | head -n 1
+}