+++ test/it/shellutils/posixer_test.sh
@@ -1,0 +1,3 @@
+TestPosixerHeader() {
+    posixer | head -n 1
+}